+++ ShapeLib/Parameters/Validation/ValidGreaterThan.cs
@@ -4,14 +4,14 @@ namespace ShapeLib
     {
         public static readonly ValidGreaterThan Zero = new ValidGreaterThan(0);
 
-        private float _greaterThan;
-        public ValidGreaterThan(float value)
+        private double _greaterThan;
+        public ValidGreaterThan(double value)
         {
             _greaterThan = value;
         }
 
-        public bool IsValid(float value) => value > _greaterThan;
-        public void ThrowIfNotValid(float value)
+        public bool IsValid(double value) => value > _greaterThan;
+        public void ThrowIfNotValid(double value)
         {
             if(IsValid(value)) return;
             throw new ArgumentException($"Parameter must be greater than {_greaterThan}, value was {value}");
